hacky implementation of model loading and rendering

+++ src/render.h
@@ -19,4 +19,4 @@ typedef union {
     u8 vals[4];
 } ColorRGBA;
 
-void UpdateAndRender(u8 *frame_buffer, u32 frame_buffer_len, i32 width, i32 height, GameState *state);
+void UpdateAndRender(u8 *frame_buffer, u64 frame_buffer_len, u32 width, u32 height, GameState *state);
